backends/requirements.txt
sqlalchemy==1.4.26
python-dotenv==0.19.1
uvicorn==0.15.0
fastapi==0.70.0
pymysql==1.0.2
pydantic==1.7.3